fix(index): require chromedriver so path is available for webdriver
See error when running `node index.js`:

```
Error: The ChromeDriver could not be found on the current PATH. Please download the latest version of the ChromeDriver from http://chromedriver.storage.googleapis.com/index.html and ensure it can be found on your PATH.
```

This error started appearing when chromedriver did not support
new Chrome version. It turns out that chromedriver was globally
installed via brew cask and now we are relying on npm chromedriver.